(#17794) remove :kernel confinement of {zfs,zpool}_version
As ZFS is now available on more or less every Solaris and BSD variant,
as well as on Linux (via FUSE), it doesn't make sense to confine these
facts to an evergrowing list of kernels.

This patch:
 - removes every reference to :kernel, as well as the SunOS/FreeBSD
   subsections from the tests
 - runs the "setcode" block only if the zfs/zpool commands are present
   on the system, and adds a test for this case
 - adds sample output of zfs/zpool on Linux (FUSE) to fixtures + a test
   using them
 - removes the "should not run on Linux" test.

The zfs/zpool providers in puppet have been recently updated in a
similar fashion (55f953bc).