(#17854) Allow install.rb to execute from outside of source root
Previously if install.rb was invoked when outside of the source root, all of
the globs would fail to find bins, libs, etc, or worse if run from / it will
find the binaries in /bin and copy them to bindir. This commit addresses that
by running the install after cding into the directory containing the install.rb
file.