Add deprecation warning for --tests flag in install.rb
The tests flag does nothing in facter and has never been functional. This
commit removes some of the points where tests is set and also adds a
deprecation warning indicating that the tests flags will be removed in the
future. It also changes the default from on to off for InstallOptions.tests.