(#17808) Add MTU facts for HP-UX
This patch extends the hack that was used to add MAC address
facts on HP-UX.  The patch also documents the hack to make
the code slightly easier to understand.

Unit tests were provided in the patch for (#17487) and are
simply uncommented by the present patch.

All of these hacks can be removed by implementation of
refactor (#17710).