(#17794) Specify how zfs facts behave when tools are configured
Without this patch applied we don't have an example of how the
zfs related facts behave when the client tools are configured after the
facts have been defined and loaded but before they've been resolved most
recently.

The expectation expressed by this patch is that the facts will resolve
when the client tools are available and will not resolve if the client
tools are not available.  Installing the client tools should be
sufficient to cause the facts to resolve.

This patch also changes the implementation to make the expected behavior
of refreshing the fact values using the code specified in the setcode
block.

Squash into