(#17917) Allow case equality confinement matching
Without this patch applied the confine functionality is limited to
testing equality.  This is a problem because some facts have different,
but similar, values over time.  For example, the use case motivating
this feature is:

    I have some facts I want to run on all generations of dell machines,
    but the manufacturer fact has changed between them so I want to be
    able to use something like /.*Dell.*/

This patch implements case equality which provides a superset of the
regular expression matching that addresses the problem.

This patch also cleans up the spec tests around confines a bit.