(Maint) Add YARD doc for get_interface_value
Without this commit there is no clear documentation for the
Facter::Util::IP.get_interface_value class method.  This is a problem
because the method behaves unexpectedly in some situations and is not
intended for public API use.

This patch fixes the problem by clearly documenting the method and
marking it as a private API.