(#17612) Follow Google docs to populate virtual => gce
Without this patch applied Facter uses lspci to determine if the system
is running in the Google Compute Engine virtual environment.  This is a
problem because Google recommends parsing the output of
/sys/firmware/dmi/entries/1-0/raw as per [Detecting if You Are Running
in Google Compute
Engine](https://developers.google.com/compute/docs/instances#dmi).
`lspci` may be unreliable over time.

This patch addresses the problem by adding a higher weight fact that
scans `/sys/firmware/dmi/entries/1-0/raw` as recommended by Google.  The
fact is confined to Linux and disables itself if
`/sys/firmware/dmi/entries/1-0/raw` does not exist in the filesystem.