(#15001) Add spec examples for ipaddress fact
Without this patch we don't have an example behaviors for the ipaddress
fact.  This patch addresses the problem by adding two examples for the
facts parsing the output of ifconfig.  These examples slightly refactor
the ipaddress fact confined to Linux in order to take advantage of the
stubbed method returning the output of the ifconfig command.

Fixture data for ifconfig output is included for Ubuntu 12.04 and Fedora
17.  The notable difference with Fedora 17 is that Net Tools 1.60
removes the "addr" substring from the interface description.