(maint) Add Travis CI Support
Without this patch facter has Travis CI configuration files, but they
don't seem to completely specify the dependency versions and the build
matrix.  This patch addresses the problem by putting the dependency
information in the conventional Gemfile location.

This patch should coincide with enabling Travis CI support for pull
requests.  A build status image is also included in the project README.
Finally, this patch also removes the Gemfile.lock since this shouldn't
be included in the repository as Facter is more of application than it
is a library.  This reverses the decision to include it, which was my
mistake, in 0b49eae and (#15464).