(maint) Add Travis CI support to active branches
Without this patch all active branches do not have proper Travis
configuration files.  This is a problem because Travis will exercise any
branch pushed to the puppetlabs organization repository on Github.  The
default behavior is to email notifications which is undesirable.

To address this problem Travis configuration settings are added to an
active branch.

Active branches are those that have recent activity when compared
against the following command:

    for k in `git branch -r |perl -pe s/^..//`
    do
      echo -e `git show --pretty=format:"%Cgreen%ci %Cblue%cr%Creset" $k|head -n 1`\\t$k;
    done | sort -r | grep origin

Which should produce output like:

    2013-01-06 11:01:40 -0800 3 hours ago   origin/1.6.x
    2013-01-05 23:33:48 -0800 14 hours ago  origin/master
    2012-12-27 10:03:43 -0800 10 days ago   origin/2.x
    2012-12-27 10:02:58 -0800 10 days ago   origin/1.7.x