This commit adds Facter::Util::IP::Ipaddress and refactors as little as possible in order to switch the ipaddress and ipaddress6 facts over to using the new module.
In order to do this I had to significantly mangle some tests and remove
some that no longer made sense.

In Facter::Util::IP I added some helper functions to search maps and
tweaked the supported_platforms function to be usable in other cases.

There's a lot more work to be done but this was the minimal amount of
changes I could make to get just the ipaddress portion of this rewrite
done.

As a bonus this allows facter to work on Linux boxes that only have
/sbin/ip instead of /sbin/ifconfig which is a nice bonus.