Maint: Ensure we don't read from /sys/class/net
Previously, the ip_spec test relied on the /sys/class/net file not
existing, so that it would flow into the HP-UX logic. But when the test
is executed on linux, the file exists, so the test short-circuits and
never evaluate the HP-UX part.

Since the facter tests are so fragile, I `expect` more than `stub` to
ensure the code takes the route I'm trying to test.