(maint) Fix physicalprocessorcount with multiple cores
Without this patch applied the physicalprocessorcount fact reports the
wrong number of processors.  Instead of reporting the number of physical
processing units, it reports the number of cores.  This patch addresses
the problem by changing the behavior to count only the number of unique
processor identities as suggested by Stefan Schulte.

We didn't catch this issue while rebasing GH-266 in GH-388 because this
behavior had no specified expectations.  This patch adds an example of
the expected behavior when there are multiple cores per processor.