Fix stupid typo.
That explains why that one spec test was failing. I assumed it was
because I wasn't on an Apple. I must've done that accidentally before I
even ran the suite, because it failed the first time, and so I assumed
that it was always failing. All spec tests pass now.