(maint) Add Ruby 2.0.0 to Travis build matrix
Without this patch we're not testing against Ruby 2.0.0 which has
recently been released.  This is a problem because we'd like a way to be
notified if a change set breaks compatibility with future supported
versions of Ruby.

This patch should not be taken as an indication that we fully support
Ruby 2.0, just as an indication that we plan to in the future.