(maint) Fix Solaris 10 spec failures
Without this patch applied we're seeing failures in the Solaris 10 spec
tests.  The root cause of this problem is a stub method that affects too
much of the system as a whole rather than the unit being tested.

This patch addresses the problem by narrowing the scope of the method
stub to limit the overall effect on the system.