(maint) Fix failing manufacturer spec on Solaris 10
Without this patch the specs are failing against solaris 10.  The root
cause of the problem is that the manufacturer spec is getting a value
from a different resolver than the one being tested.  The example
expected the result to be nil rather than being resolved by the
alternative resolver.

This patch fixes the issue by changing the expectation such that the
manufacturer spec is not the value set by the resolver being tested.