[packaging] Update mocks for rpmbuilder mock format
The Puppet Labs mocks created by rpmbuilder have assumed a new format, pl-el-*
vs el-*, in order to avoid overwriting the configurations supplied by the mock
package. This commit updates the mocks in facter to reflect the new standard,
so that we can continue to build packages with builders created with the
rpmbuilder module.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>