Replace RDoc::Usage with Trollop
As per original work by Ashley Penney <apenney@gmail.com>,
RDoc::Usage is gone in Ruby 1.9.x, so displaying/parsing options
have been converted to Trollop.

As per the original pull request, this removes the vendored copy of
Trollop, and replaces it with a dependecy on Trollop; this has been
reflected in the Gemfile as well as the gemspec.

This includes the relevant changes to the banner, but does not
include the RDoc changes to bin/facter (https://github.com/apenney/facter/commit/a4fa9121971e5c7d08be93659b64bb3ffc89d01e)