(maint) Update Gemfile to work well on windows
Without this patch applied, the following issues are encountered trying
to get Facter working in a development environment on Windows.  This is
with Ruby 1.9.3 and Bundler 1.2.4.

First, the runtime provided by rubyinstaller.org is unable to verify the
SSL certificate of rubygems.org.  For Windows, we work around the
problem by allowing the user to specify an alternate gem source using
the GEM_SOURCE environment variable.  For example,
`$ENV:GEM_SOURCE="http://rubygems.org"`.

Next, `bundle exec facter --version` does not work because Facter 1.7
depends on the win32-dir gem and this dependency is not specified in the
Gemfile.  This patch addresses the problem by specifying the dependency
on win32-dir.