(maint) Clean up require => false in Gemfile
This is just a maintenance patch to clean up all of the `:require =>
false` entries in the bundler Gemfile.  These statements prevent Bundler
from loading the gem libraries when running `bundle exec` or
`Bundler.setup`.  We don't run `Bundler.setup` in the code, so the
libraries will only be pre-loaded when run through `bundle exec` which
is by design.

This does not appear to cause any problems with the spec examples.