(#16948) Fix hardwaremodel fact on windows 32 bit
Without this patch applied the hardwaremodel fact will return 'x64' when
a 32 bit operating system is running inside of a 64 bit CPU.  This is a
problem because it is inconsistent with the way Facter operates on other
platforms where the behavior is to inspect the operating system rather
than the CPU itself.

This patch fixes the problem by checking the AddressWidth attribute
which will report 32 when a 32 bit OS is running on a 64 bit CPU.  The
attribute will report 64 when a 64 bit OS is running.