(maint) Add pry to Gemfile as a development dependency
Without this patch it's difficult to do interactive debugging and
introspection of a live, running facter process.  This is a problem
because it makes developing Facter more difficult than it needs to be.
This patch addresses the problem by listing pry as a development
dependency of Facter.  With this patch applied, the following command
may be used to drop into a debugger using the `require 'pry';
binding.pry` statement while running the spec tests.

    bundle exec rspec spec/

This patch also adds the yard and redcarpet development dependencies
which synchronize Facter 1.7 with Puppet master regarding YARD
documentation.