Add on_flush fact resolution DSL method
Without this patch it is difficult to model an expensive system call
inside of an object instance, generate N dynamic facts from the
data contained in the model, then keep all of the resulting fact values
in sync with one another without making N system calls.

This patch addresses the problem by implementing a callback mechanism
when fact values are flushed.  Fact authors may define a so-called
`on_flush` block that will be executed when the fact value is flushed by
Facter.

This block is then useful to invalidate the model data referenced by all
of the N dynamic facts.  The first fact to refresh it's value will then
incur the expense of the system call, but the next N-1 facts do not need
to make the expensive system call as their data has already been
refreshed.

That is to say, flushing the value for all N facts will only incur 1
system call instead of N system calls using the traditional setcode
block.