(#19761) Added virtual and is_virtual fact on Windows
Without this patch, the virtual and is_virtual fact is not set
on Windows computers. This patch fixes the problem, giving an
initial coverage for some hypervisors.

The solution proposed is based on the manufacturer and model
properties from WMI object Win32_ComputerSystem. This first version
is able to detect the following systems: HyperV, VirtualBox, Vmware
and KVM.