(#19989) Filter virt-what warnings from virtual fact
Presently, the virt-what command puts errors on stdout, prefixed with
virt-what:.  This ends up in the virtual fact, which is undesirable.

When a fixed virt-what is released and available for supported systems,
the 'output.gsub /^virt-what: .*$/' can be removed from
Facter::Util::Virtual::virt_what().

A virt-what fix was committed upstream¹.

Thanks to Adrien Thebo for help on making this testable.

¹ http://git.annexia.org/?p=virt-what.git;a=commitdiff;h=2f47e06