(packaging) Add dependency on virt-what for rpm and deb packages
In commit cf43fc0092f0476d378ea05dd8e21fc170a51bdf (#8210), virt-what was added
as a utility for better virtualization detection. Without virt-what installed,
that better detection is not available, so this commit adds virt-what as a
dependency for debian and rpm packages. virt-what is a lightweight package
(~25k), that is available on all currently supported flavors of debian and rpm
based systems.