(maint) Fix unhandled exception when zoneadm is missing
Without this patch the puppet master spec examples are currently failing
when run on Solaris 10.  The root cause of this failure is an assumption
that the `zoneadm` command is always available on Solaris 10 systems.

Because of this assumption, the method that parses out the zone
information tries to call split on an instance variable that is assumed
to be a string, but is actually `nil` when the `zoneadm` command is
missing.

This patch addresses the problem by checking if the instance variable is
truthy and only then calling the split method on it.  Otherwise a debug
message records the fact that the zoneadm command is not available which
causes the zone facts to be unset.