(doc) Update the Resolution#confine API docs
Without this patch, the API documentation for the
Facter::Util::Resolution#confine is a bit muddled and confusing.  This
confusion is caused by the different, overloaded, forms of arguments
that #confine accepts.

This patch addresses the problem by using the @overload YARD tag to
separate out each form, document the arguments for each form, and
provide a clear example for each form.