(#14522) Force encoding of data read from /proc/self/status
Without this patch the previous fix for this issue worked well, but was
actually testing and fixing a different problem than described in the
issue.  Reading the /proc/self/status file on Solaris works well,
however the file contents are very different from the Linux platform.
On linux the file contains US-ASCII encoded text which is fully
compatible with our default UTF-8 encoding.  On Solaris, however, the
file is raw data representing the memory space of the process.  This raw
data has no encoding, and as such, causes the invalid byte sequence when
compared with a regular expression.

This patch addresses the problem by forcing the encoding of the string
into UTF-8, removing any invalid characters.  The comparison is not a
problem on MRI versions which do not have `String#encode!`.