Add the lsbminordistrelease fact
Without this patch there is not an easy way to get the minor versions
from the LSB Distribution Release version without resorting to using
split or other such functions.

This patch addresses the problem by adding a new fact,
`lsbminordistrelease` which directly contains the minor version parsed
out of the full lsbdistrelease version.

This fact will be defined whenever the lsbdistrelease fact is defined.