(maint) Fix Facter::Util::IP.exec_ifconfig examples
Without this patch the Facter::Util::IP.exec_ifconfig examples only set
stub methods, they don't actually set expectations that ifconfig is
executed as the examples describe.

This patch addresses the problem by updating the examples to reflect the
intent of the behavior expressed in the description of the example.