(#20236) Add example that reproduces dmidecode issue
Without this patch we don't have an example that reproduces the issue
described in 20236.  This is a problem because we don't automatically
know when we've fixed the issue.

This patch addresses the problem by adding a single example that
reproduces the issue described in 20236; dmidecode is not consulted when
the lspci command returns data but does not result in a determination of
the virtual platform.