(#20229) Fix per-interface netmask_* facts with net-tools 1.60
Without this patch the interface specific facts for the netmask are not
correctly parsed on systems with net-tools 1.60 or later, such as
Archlinux.

This patch addresses the problem by adjusting the regular expression to
take into account the new format of the ifconfig command.

This patch is specific to the netmask_* facts and does not affect the
behavior of the general `netmask` fact.