(#20229) Fix netmask fact with Linux net-tools 1.60
Without this patch the netmask fact does not return a value on Archlinux
because of the new format of the net-tools package.

This patch addresses the problem by updating the regular expressions for
the netmask fact.  Both the old format and the new format are covered by
the regular expression now.