(#7643) Export the Win32_OperatingSystem WMI/CIM class
Without this patch, there is a large set of data available in Windows that
users do not have the ability to make decisions on.

This patch adds exposes all the properties in the singleton
Win32_OperatingSystem WMI/CIM class as facts. The facts will have
'wmi_win32_operatingsystem_' prepended. The data received from the
WMI/CIM class is exposed raw in the facts. This is actually preferable
as users can refer to MSDN documentation for the class to understand the
purpose of different values.