(#20321) Be more descriptive in deprecation message
Facter::Util::Resoultion.exec can be used to run arbitrary commands. On
windows it is currently allowed to pass a shell builtin but we raise a
deprecation warning.

But the deprecation warning did not include the actual commandline it is
complaining about so if you have custom facts that make use of this
feature you may ask yourself why you see the message.

Add the actual command to the deprecation message so it easier to
resolve the issue.