(maint) Make facter gemspec report correct version
Without this patch the Gem system thinks that the version of Facter
running from a development branch is `1.6.11` even though many versions
have been released since then.  This is a problem because a project that
requires Facter >= 1.7.0 should be able to satisfy this dependency using
the `stable` development branch, however it cannot because the version
information is out of date.

This patch addresses the problem by using Facter's public version API to
determine the version of Facter contained in the git branch.

This patch also attempts to make it clear to the reader that the gemspec
is intended for use with the VCS branches under active development and
_not_ as the way to produce a gem release package.