(maint) Silence no facts loaded warning in specs
Without this patch the mocked fact value in some of the tests are
producing a warning that no facts are loaded.  This is a problem because
the warning is a red herring side effect of the mocking itself.

This patch addresses the problem by stubbing the Facter.warnonce method
for these specific examples that have no facts loaded at the time of
setting the expectation.