(maint) Replace rspec >= 2.11 expect(foo).to with foo.should
Without this patch the specs are failing in our CI system because we
have rspec 2.9 installed and the change set in c028d61 introduced a lot
of examples that use the `expect().to` syntax added in rspec-2.11.

This patch addresses the problem by replacing all instances of
`expect(foo).to` with `foo.should`.