(FACT-197) Backport Solaris processorcount fix
A fix for redmine #18215 was comitted to master as commits
db87076..4c035e8. This squashes those commits and rebases them to the
stable branch as a bugfix for the next facter release. Commit messages
from the original commits follow

The first commit's message is:
Refined logic for SunOS processorCount

Refined logic to use kstat only on SunOS < 5.8, psrinfo on everything
from 5.8 and above.

This is the 2nd commit message:

(#18215) Updated processor_spec.rb for Solaris changes

Updated processor_spec.rb for Solaris changes to support fixes for
issue #18215.

This is the 3rd commit message:

processorcount spec wrongly used physicalprocessorcount fact

This is the 4th commit message:

adding psrinfo from Solaris with 24 CPUs and using it in processor spec.

This is the 5th commit message:

removing double quotes in a spec

This is the 6th commit message:

removed duplicate tests for Solaris

Conflicts:
	lib/facter/processor.rb
	spec/unit/processor_spec.rb