(#20915) Read dmi entries as binary data
The DMI entry information that is read to determine the virtual platform
is actually binary data. When it is interpreted as an encoded string,
then there will be encoding errors. This changes it to be read in "rb"
mode, which has nearly no effect on ruby 1.8.7, but causes it to be read
as Encoding::ASCII_8BIT on ruby 1.9.3 (which is how ruby deals with
binary data).