(maint) remove unused code from resolution_spec
Prior to this commit, the resolution_spec.rb file tests setcode for both
a windows and non-windows system. The tests come out exactly the same
for both systems. An investigation into the code suggests that this
platform distinction is never actually used. This commit removes the
part of the test that specifies what to do on a windows vs. non-windows
system.