(maint) remove whitespace and alter formatting
Prior to this commit there were some whitespace errors in the code,
relevant information for the test_cases was stored in two different data
sets, and there was unnecessary comments. This commit fixes the
whitespace issues, reformats the datastructure of the test_cases hash,
and gets rid of the comments.