(maint) Add acceptance test for stderr facter output
Facter often has facts which don't correctly capture or suppress stderr,
allowing it to be displayed during the facter run. This can cause problems with
either facter runs through cron, or if puppet is running through cron, puppet
will also throw the same output to stderr. This commit adds a simple acceptance
test that runs facter and verifies that the stderr stream was empty after the
run to avoid this problem.