Don't build debuginfo rpm package
Prior to this commit, mock would spit out a debuginfo package on facter.
This is pointless as facter contains zero compiled code. It would be a
noarch package, save for some arch-specific dependencies. So, since I
can't fix that bug in RPM, I can make it so I don't have to remove the
debuginfo package every single time.

Signed-off-by: Michael Stahnke <stahnma@puppetlabs.com>