(FACT-341) - Added operatingsystemrelease support to Windows
Without this patch, the fact 'operatingsystemrelease' was set to
kernel version fact for all the windows flavors. This patchs brings
the option of having the common release name for the windows
machines, something that can help in writing specific puppet
manifests and other tasks according to the specific windows version.

The patch is based on a WMI call to Win32_OperatingSystem. This
aproach has some drawbacks but I think the code is simple and
easy to mantain and update. Anyway, extracting the windows version
is not an easy task for Microsoft as well:
http://msdn.microsoft.com/en-us/library/windows/desktop/ms724429%28v=vs.85%29.aspx

The original Redmine issue is (#7621)

Conflicts:
	lib/facter/operatingsystemrelease.rb