(maint) Stub extraneous facts when testing ifconfig
If the higher weight ifconfig fact fails to resolve, Facter will try to
determine the IP address by resolving the hostname. When testing to
ensure that the ifconfig based resolution returns nil, the resolution
fact will be called and return an unexpected value. This commit stubs
the hostname fact when testing the ifconfig ipaddress resolution to
ensure the tests behave consistently.